+++ bpm-engine-impl/src/main/java/io/takari/bpm/AbstractEngine.java
@@ -11,6 +11,7 @@ import io.takari.bpm.event.EventPersistenceManager;
 import io.takari.bpm.lock.LockManager;
 import io.takari.bpm.persistence.PersistenceManager;
 import io.takari.bpm.planner.Planner;
+import io.takari.bpm.resource.ResourceResolver;
 import io.takari.bpm.state.*;
 import io.takari.bpm.task.UserTaskHandler;
 import org.slf4j.Logger;
@@ -42,8 +43,6 @@ public abstract class AbstractEngine implements Engine {
     protected abstract EventPersistenceManager getEventManager();
 
     protected abstract LockManager getLockManager();
-
-    protected abstract UserTaskHandler getUserTaskHandler();
 
     protected abstract Configuration getConfiguration();