+++ bpm-engine-impl/src/main/java/io/takari/bpm/resource/ClassPathResourceResolver.java
@@ -1,0 +1,12 @@
+package io.takari.bpm.resource;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public class ClassPathResourceResolver implements ResourceResolver {
+
+    @Override
+    public InputStream getResourceAsStream(String name) throws IOException {
+        return ClassLoader.getSystemResourceAsStream(name);
+    }
+}