+++ bpm-engine-impl/src/main/java/io/takari/bpm/resource/ResourceResolver.java
@@ -1,0 +1,9 @@
+package io.takari.bpm.resource;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public interface ResourceResolver {
+
+    InputStream getResourceAsStream(String name) throws IOException;
+}