+++ bpm-engine-impl/src/test/java/io/takari/bpm/EngineHolder.java
@@ -15,6 +15,8 @@ import io.takari.bpm.persistence.InMemPersistenceManager;
 import io.takari.bpm.persistence.PersistenceManager;
 import io.takari.bpm.planner.DefaultPlanner;
 import io.takari.bpm.planner.Planner;
+import io.takari.bpm.resource.ClassPathResourceResolver;
+import io.takari.bpm.resource.ResourceResolver;
 import io.takari.bpm.task.ServiceTaskRegistryImpl;
 import io.takari.bpm.task.UserTaskHandler;
 import org.slf4j.Logger;
@@ -48,6 +50,7 @@ public class EngineHolder {
     private final LockManager lockManager;
     private final Configuration configuration;
     private final DelegatingUserTaskHandler userTaskHandler;
+    private final ResourceResolver resourceResolver;
 
     public EngineHolder() throws Exception {
         serviceTaskRegistry = new ServiceTaskRegistryImpl();
@@ -61,9 +64,10 @@ public class EngineHolder {
         configuration = new Configuration();
         lockManager = new SingleLockManagerImpl();
         userTaskHandler = new DelegatingUserTaskHandler();
+        resourceResolver = new ClassPathResourceResolver();
         executor = wrap(new DefaultExecutor(configuration, expressionManager, Executors.newCachedThreadPool(),
                 interceptorHolder, indexedProcessDefinitionProvider, uuidGenerator, eventManager, persistenceManager,
-                userTaskHandler));
+                userTaskHandler, resourceResolver));
 
         engine = new AbstractEngine() {
 
@@ -118,11 +122,6 @@ public class EngineHolder {
             @Override
             public EventService getEventService() {
                 return eventService;
-            }
-
-            @Override
-            protected UserTaskHandler getUserTaskHandler() {
-                return userTaskHandler;
             }
         };